requirements.txt
torch==2.2.1
torchvision==0.17.1
matplotlib==3.8.3
lightning==2.2.1
h5py==3.10.0
scipy==1.12.0
nltk==3.8.1
pytest==8.1.1
torchmetrics==1.3.1
wandb==0.16.4
jsonargparse==4.27.6
tensorboard==2.16.2
hydra-core==1.3.2
omegaconf==2.3.0
google-api-python-client==2.123.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
toml==0.10.2
